feat: add merchant bill

+++ tests/domain/account_test.go
@@ -34,7 +34,7 @@ func TestFlowAccount(t *testing.T) {
 		t.Error(err)
 		t.FailNow()
 	}
-	nowBalance := repo.GetMember(m.GetAggregateRootId()).GetAccount().GetValue().FlowBalance
+	nowBalance := repo.GetMember(int64(m.GetAggregateRootId())).GetAccount().GetValue().FlowBalance
 	if nowBalance != balance {
 		t.Logf("before:%.2f  now:%.2f", float64(balance)/100, float64(nowBalance)/100)
 		t.FailNow()
